tests: the inference tests split by subject, one file per reading budget
src/hir/typeinfer/tests.rs held 899 lines across five unrelated subjects, so
reading it end to end — which is what lets somebody edit it — cost more than
the subject they came for.

One module each: narrow for what a guard or a dispatch arm proves, rettype
for what an op's result proves, prune for dead-arm removal, contract for
prove-or-reject, monomorphize for wrapper collapse. tests.rs keeps the three
helpers they all compile through. No test changed.
